redesign the site header
the navigation links in the header are now always visible rather than
inside a hamburger menu. the links are now styled with a grey block
background that changes to yellow on hover. the link to the active
section of the site will always have a yellow background, and the others
grey. most pages, but not all, fall within a section listed in the nav
links, so most of the time one of the links is yellow.

the site logo is a new SVG I created using D3 in an Observable notebook
- https://observablehq.com/d/957a7c4a03bfce94 - it is based on an
  Interrupted Sinusoidal projection.