redesign the site footer
the footer now has a yellow background, the full width of the viewport.
the navigation links have arrow bullet points. the RSS link has been
separated from the nav links and has an RSS icon next to it. all text
has been styled (near) black. the link hover styles vary between nav
links, the RSS link, the attribution links, and the changelog link. the
RSS link *would* have been styled the same as the nav links, if not for
horrible "tricks" to position it where it is in relation to the RSS
icon...