redesign the homepage recent feed
the recent posts feed on the homepage is now comprised of two columns.
the first lists blog posts and the second lists weeknotes. each column
has its own heading.

the items on each list are now in "compact" form: only a link to the
post, no preview/summary text or metadata.

the "recent" and "explore" headings have been removed. the "explore
buttons" are now the only thing between the recent feed and the tag
cloud. there are now 3 buttons, the first for viewing all blogs, the
second for viewing all weeknotes, and the third for viewing "everything"
which is currently a list of all blogs and weeknotes, no other site
sections are included currently. new index pages were created to
facilitate this revamp of the buttons.